+++ packages/react-web-kit/src/components/molecules/DatePicker/DatePicker.tsx
@@ -1,5 +1,4 @@
 import {
-  Button,
   Calendar,
   DatePicker as DatePickerCore,
   DatePickerProps,
@@ -7,6 +6,8 @@ import {
   Masks,
   SelectionType,
   CalendarIcon,
+  PressableSurface,
+  useTheme,
 } from '@tecsinapse/react-core';
 import React, { useCallback, useMemo, useState } from 'react';
 import { Dropdown } from '../../atoms/Dropdown';
@@ -21,6 +22,7 @@ export type WebDatePickerProps<T extends SelectionType> = Omit<
   | 'requestShowCalendar'
 > & {
   callbackAfterValidated?: (valid: boolean, message?: string) => void;
+  invalidDateLabel?: string;
 };
 
 export const DatePicker = <T extends SelectionType>({
@@ -29,6 +31,10 @@ export const DatePicker = <T extends SelectionType>({
   locale,
   onChange,
   callbackAfterValidated,
+  placeholder,
+  disabled,
+  label,
+  invalidDateLabel = 'Invalid date',
   ...rest
 }: WebDatePickerProps<T>): JSX.Element => {
   const [visible, setVisible] = useState(false);
@@ -36,6 +42,7 @@ export const DatePicker = <T extends SelectionType>({
   const show = useCallback(() => setVisible(true), []);
   const close = useCallback(() => setVisible(false), []);
   const [error, setError] = useState<boolean>(false);
+  const theme = useTheme();
 
   const getYear = useMemo(() => {
     if (value) {
@@ -62,9 +69,9 @@ export const DatePicker = <T extends SelectionType>({
   }, [value]);
 
   const checksFullRange = useCallback(() => {
-    if(type === 'range' && !(value as DateRange)?.highest)
-      onChange?.(undefined)
-  }, [value])
+    if (type === 'range' && !(value as DateRange)?.highest)
+      onChange?.(undefined);
+  }, [value]);
 
   const controlComponent = (onPress, displayValue) => {
     return (
@@ -75,14 +82,14 @@ export const DatePicker = <T extends SelectionType>({
             (controlledInput ?? []).length > 0
           ) {
             setError(true);
-            callbackAfterValidated?.(false, 'Data inválida');
+            callbackAfterValidated?.(false, invalidDateLabel);
           }
           if (controlledInput?.length === 8) {
             const auxData = parse(controlledInput, 'ddMMyyyy', new Date(), {
               locale,
             });
             const isValidDate = isValid(auxData);
-            callbackAfterValidated?.(isValidDate, 'Data inválida');
+            callbackAfterValidated?.(isValidDate, invalidDateLabel);
 
             if (isValidDate && auxData !== value) {
               setError(false);
@@ -91,12 +98,13 @@ export const DatePicker = <T extends SelectionType>({
           }
           if (controlledInput?.length === 0) {
             setError(true);
-            callbackAfterValidated?.(false, 'Data inválida');
+            callbackAfterValidated?.(false, invalidDateLabel);
           }
         }}
+        disabled={disabled}
         mask={Masks.DATE}
-        value={displayValue}
-        hint={error ? 'Data inválida' : undefined}
+        value={displayValue ?? ''}
+        hint={error ? invalidDateLabel : undefined}
         variant={error ? 'error' : 'default'}
         onChange={input => {
           setControlledInput(input);
@@ -105,16 +113,20 @@ export const DatePicker = <T extends SelectionType>({
             setError(false);
           }
         }}
-        placeholder={'Não informada'}
+        placeholder={placeholder}
+        label={label}
+        placeholderTextColor={
+          disabled ? theme.color.secondary.light : undefined
+        }
         rightComponent={
-          <Button
+          <PressableSurface
             effect={'none'}
-            variant={'text'}
             style={{ padding: 0 }}
             onPress={onPress}
+            disabled={disabled}
           >
             <CalendarIcon name="calendar-sharp" type="ionicon" size="centi" />
-          </Button>
+          </PressableSurface>
         }
       />
     );
@@ -130,6 +142,7 @@ export const DatePicker = <T extends SelectionType>({
         value={value}
         type={type}
         year={getYear}
+        disabled={disabled}
         format={'dd/MM/yyyy'}
         month={getMonth}
         requestShowCalendar={show}
@@ -152,20 +165,25 @@ export const DatePicker = <T extends SelectionType>({
         value={value}
         type={type}
         year={getYear}
+        disabled={disabled}
         format={'dd/MM/yyyy'}
         month={getMonth}
         requestShowCalendar={show}
         requestCloseCalendar={close}
-        renderCalendar={(calendar, handleBlur) =>
-                  <Dropdown visible={visible} setVisible={setVisible}
-                            onClickAway={() => {
-                                handleBlur?.()
-                                checksFullRange()
-                            }}
-                  >
-                    {calendar}
-                  </Dropdown>
-        }
+        placeholder={placeholder}
+        label={label}
+        renderCalendar={(calendar, handleBlur) => (
+          <Dropdown
+            visible={visible}
+            setVisible={setVisible}
+            onClickAway={() => {
+              handleBlur?.();
+              checksFullRange();
+            }}
+          >
+            {calendar}
+          </Dropdown>
+        )}
       />
     );
   }